% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{.check_oracle_resourcer_prereqs}
\alias{.check_oracle_resourcer_prereqs}
\title{Check oracle.resourcer prerequisites}
\usage{
.check_oracle_resourcer_prereqs()
}
\value{
A message indicating whether the required package is installed.
}
\description{
This function checks if the required package \code{ROracle} is installed.
If the required package is missing, it stops and returns an error message.
If the package is installed, it returns a confirmation message.
}
\details{
Please note that this function is added to the package so that, when registered as a method,
the Opal UI correctly displays the \code{oracle.resourcer} package in the list of installed packages,
facilitating its management and discoverability by the data manager.
}
