% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OracleResourceConnector.R
\docType{class}
\name{OracleResourceConnector}
\alias{OracleResourceConnector}
\title{Oracle DBI resource connector}
\format{
A R6 object of class OracleResourceConnector
}
\description{
Oracle DBI resource connector

Oracle DBI resource connector
}
\details{
Makes a Oracle DBI connection from a resource description.
}
\section{Super class}{
\code{\link[resourcer:DBIResourceConnector]{resourcer::DBIResourceConnector}} -> \code{OracleResourceConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OracleResourceConnector-new}{\code{OracleResourceConnector$new()}}
\item \href{#method-OracleResourceConnector-isFor}{\code{OracleResourceConnector$isFor()}}
\item \href{#method-OracleResourceConnector-createDBIConnection}{\code{OracleResourceConnector$createDBIConnection()}}
\item \href{#method-OracleResourceConnector-clone}{\code{OracleResourceConnector$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="DBIResourceConnector" data-id="closeDBIConnection"><a href='../../resourcer/html/DBIResourceConnector.html#method-DBIResourceConnector-closeDBIConnection'><code>resourcer::DBIResourceConnector$closeDBIConnection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="DBIResourceConnector" data-id="getTableName"><a href='../../resourcer/html/DBIResourceConnector.html#method-DBIResourceConnector-getTableName'><code>resourcer::DBIResourceConnector$getTableName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="DBIResourceConnector" data-id="readDBTable"><a href='../../resourcer/html/DBIResourceConnector.html#method-DBIResourceConnector-readDBTable'><code>resourcer::DBIResourceConnector$readDBTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="DBIResourceConnector" data-id="readDBTibble"><a href='../../resourcer/html/DBIResourceConnector.html#method-DBIResourceConnector-readDBTibble'><code>resourcer::DBIResourceConnector$readDBTibble()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OracleResourceConnector-new"></a>}}
\if{latex}{\out{\hypertarget{method-OracleResourceConnector-new}{}}}
\subsection{Method \code{new()}}{
Creates a new OracleResourceConnector instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OracleResourceConnector$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A OracleResourceConnector object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OracleResourceConnector-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-OracleResourceConnector-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check that the provided resource has a URL that locates a Oracle object: the URL scheme must be "oracle".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OracleResourceConnector$isFor(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource object to validate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OracleResourceConnector-createDBIConnection"></a>}}
\if{latex}{\out{\hypertarget{method-OracleResourceConnector-createDBIConnection}{}}}
\subsection{Method \code{createDBIConnection()}}{
Creates a DBI connection object from a resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OracleResourceConnector$createDBIConnection(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A DBI connection object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OracleResourceConnector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OracleResourceConnector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OracleResourceConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
